Possibility to add a local key
It is now possible to define a key in ~/.passtools.rc, this key should be a
long random string. Using it make your master password insufficient to
retrieve all your credentials.

The --no-key|-K parameter can be used to ignore the key defined in rc file
while generating the password, this parameter should ease password migration.